UserspaceEmulator: Support munmap/mprotect with partial mappings
Fixes #5663.

+++ Userland/DevTools/UserspaceEmulator/Emulator.h
@@ -99,7 +99,7 @@ private:
     u32 virt$mmap(u32);
     FlatPtr virt$mremap(FlatPtr);
     u32 virt$mount(u32);
-    u32 virt$munmap(FlatPtr address, u32 size);
+    u32 virt$munmap(FlatPtr address, size_t size);
     u32 virt$gettid();
     u32 virt$getpid();
     u32 virt$unveil(u32);
